feat: 加了日志，加了一些测试。问题测试没过，似乎还是本机的数据库稳定一点，调试难度也少一点
