+++ Cargo.toml
@@ -7,5 +7,7 @@ edition = "2021"
 
 [dependencies]
 tokio-postgres = "0.7.10"
-postgresql_embedded = { version = "0.9.2", features = ["tokio"] }
-tokio = { version = "1.37.0", features = ["full"] }               # async networking
+postgresql_embedded = { version = "0.9.2", features = ["tokio", "bundled"] }
+tokio = { version = "1.37.0", features = ["full"] }                          # async networking
+log = "0.4.21"
+log4rs = "1.3.0"
